refactor(console): message controller (wave 3)
Extract the native Console message-transcript cluster (35 *message*-named
methods, ~1,577 lines, the largest wave-3 cluster) into
ConsoleMessageController (tldw_chatbook/UI/Console_Modules/message.py).

29 of the 35 move (plus 7 exclusively-owned helpers whose names don't
match the *message* pattern); 6 stay on ChatScreen (DOM access, or
name-pattern false positives for the realtime/image-generation clusters
already documented as staying screen-owned by prior waves). handle_console_
message_action (294 lines, the risk centre) has no DOM access and moves
cleanly, fanning out into ~19 named constructor callables for the
change-review/image-generation/session/workspace clusters it dispatches
into.

17 of the 29 moved methods keep a thin (<=3-line) delegation on ChatScreen
under their original name: this cluster's pre-move test suite reaches an
unusually large number of these by their private name directly
(screen.handle_console_message_action(...), ChatScreen._serialize_console_
message(...), screen._select_console_message_variant(...), etc, built up
across many prior test-writing rounds) or from production callers this
wave didn't touch (on_button_pressed, _restore_native_console_state,
_serialize_native_console_state, ~10 command-dispatch call sites for
_append_native_console_system_message). 3 more moved methods had exactly
one staying caller each, fixed by a direct call-site edit instead of a
delegation. Message state (_last_console_action, _pending_console_delete_
message_id, _console_original_attempt_previews, _console_speaking_message_
id, _pending_console_swipe_selection) moved to the controller with proxy
properties kept on ChatScreen, matching every prior wave's pattern --
_console_speaking_message_id's proxy is load-bearing for console_
transcript.py's bare-name getattr(self.screen, ...) reach.

Fixed three test-helper defect-class instances the move exposed (single-
module patches/bare-screen fixtures that stopped governing the moved
path): get_cli_setting and blocked_reason/resolve_console_artifact_owner_
request monkeypatches retargeted from chat_screen's module namespace to
message's; two bare-ChatScreen.__new__ test fixtures
(Tests/Chat/test_console_generation_actions.py, Tests/UI/test_console_
native_chat_flow.py, Tests/Chat/test_console_attachment_riders.py) given a
real ConsoleMessageController construction so their AttributeError-on-
_message gap doesn't surface. active_session_is_ephemeral is wired through
ChatScreen's own existing delegation rather than straight to
ConsoleSessionController, because 4 pre-existing tests monkeypatch it at
the screen level.

chat_screen.py: 20,964 -> 19,636 lines, 612 -> 602 methods. pyflakes
clean (matches pre-move baseline exactly, only line numbers shifted).
Characterisation tests (Tests/UI/test_console_message_controller.py)
committed and run green against unmodified code before this extraction
(see prior commit). Two known-unrelated failures in the mandated
verification sweep, both reproduced identically on the pre-wave-3 base
commit (a1758e348) via a disposable worktree: a DOM-mock gap in
test_generate_image_handler_restores_draft_when_batch_raises (image-
generation cluster, untouched) and a terminal-geometry pilot-click flake
in test_save_image_button_reflects_the_real_screen_ephemeral_accessor.
The size-ratchet slack test fails as expected -- lowering the budget is
wave-3 task 4's job, not this task's.
